+++ apps/frontend/api/useGetProjectsUser.ts
@@ -1,27 +1,17 @@
 "use client";
-import { useAuth } from "@/app/(auth)/AuthProvider";
-import { SERVERL_BASE_URL } from "@/config";
 import { project } from "@bitrock/db";
 import { useCallback, useEffect, useState } from "react";
+import { fetchAllProjects } from "./server/project/fetchAllProjects";
 
 export function useGetProjectsUser() {
-  const { user, session } = useAuth();
   const [isLoading, setIsLoading] = useState(true);
   const [projects, setProjects] = useState<project[]>([]);
 
   const fetchProjects = useCallback(async () => {
-    if (!user?.id) return;
-    return fetch(`${SERVERL_BASE_URL}/projects/user/${user.id}`, {
-      method: "GET",
-      headers: {
-        "Content-Type": "application/json",
-        Authorization: `Bearer ${session?.access_token}`,
-      },
-    })
-      .then((res) => res.json())
+    return fetchAllProjects({ params: null })
       .then((data) => setProjects(data))
       .finally(() => setIsLoading(false));
-  }, [session?.access_token, user?.id]);
+  }, []);
 
   useEffect(() => {
     fetchProjects();
