+++ apps/frontend/app/dashboard/page.tsx
@@ -3,11 +3,12 @@ import { fetchUserStats } from "@/api/server/dashboard/fetchUserStats";
 import { fetchUserPermits } from "@/api/server/permit/fetchUserPermits";
 import { fetchUserTimesheet } from "@/api/server/timesheet/fetchUserTimesheet";
 import CalendarView from "@/components/dashboard/calendar-view";
-import DashboardHeader from "@/components/dashboard/dashboard-header";
+import { DashboardHeader } from "@/components/dashboard/dashboard-header";
 import DashboardSummary from "@/components/dashboard/dashboard-summary";
 import HoursChart from "@/components/dashboard/hours-chart";
 import NotificationsCard from "@/components/dashboard/notifications-card";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
+import { getUserInfoFromCookie } from "@/utils/supabase/server";
 import type { Metadata } from "next";
 
 export const metadata: Metadata = {
@@ -16,6 +17,7 @@ export const metadata: Metadata = {
 };
 
 export default async function DashboardPage() {
+  const user = await getUserInfoFromCookie();
   const summary = await fetchUserStats();
   const latestNotifications = await fetchLatestNotifications();
 
@@ -23,7 +25,7 @@ export default async function DashboardPage() {
   const permits = await fetchUserPermits();
   return (
     <div className="space-y-6">
-      <DashboardHeader />
+      <DashboardHeader user={user} />
       <DashboardSummary summary={summary} />
 
       <Tabs defaultValue="chart" className="w-full">
