+++ apps/frontend/app/progetti/page.tsx
@@ -1,5 +1,6 @@
 import ProjectsHeader from "@/components/projects/projects-header";
 import ProjectsTable from "@/components/projects/projects-table";
+import { getUserInfoFromCookie } from "@/utils/supabase/server";
 import type { Metadata } from "next";
 
 export const metadata: Metadata = {
@@ -7,10 +8,11 @@ export const metadata: Metadata = {
   description: "Gestione dei progetti aziendali",
 };
 
-export default function ProjectsPage() {
+export default async function ProjectsPage() {
+  const user = await getUserInfoFromCookie();
   return (
     <div className="space-y-6">
-      <ProjectsHeader />
+      <ProjectsHeader user={user} />
       <ProjectsTable />
     </div>
   );
