+++ apps/frontend/app/utenti/[id]/page.tsx
@@ -1,4 +1,5 @@
 import UserDetail from "@/components/users/user-detail";
+import { getUserInfoFromCookie } from "@/utils/supabase/server";
 import type { Metadata } from "next";
 
 export const metadata: Metadata = {
@@ -12,9 +13,10 @@ export default async function UserDetailPage({
   params: Promise<{ id: string }>;
 }) {
   const { id } = await params;
+  const user = await getUserInfoFromCookie();
   return (
     <div className="space-y-6 h-full w-full">
-      <UserDetail id={id} />
+      <UserDetail id={id} userData={user} />
     </div>
   );
 }
