+++ apps/frontend/app/utenti/page.tsx
@@ -1,6 +1,7 @@
 import { findUsers } from "@/api/server/user/findUsers";
 import UsersHeader from "@/components/users/users-header";
 import UsersTable from "@/components/users/users-table";
+import { getUserInfoFromCookie } from "@/utils/supabase/server";
 import { revalidatePath } from "next/cache";
 
 export default async function UsersPage({
@@ -12,6 +13,7 @@ export default async function UsersPage({
 }) {
   const { params } = await searchParams;
   const users = await findUsers(params);
+  const user = await getUserInfoFromCookie();
 
   const revalidate = async () => {
     "use server";
@@ -20,8 +22,8 @@ export default async function UsersPage({
 
   return (
     <div className="space-y-6">
-      <UsersHeader />
-      <UsersTable users={users} refetch={revalidate} />
+      <UsersHeader user={user} />
+      <UsersTable users={users} refetch={revalidate} user={user} />
     </div>
   );
 }
